Moved notebooks back to Python/examples folder.
