+++ .travis/python2.build
@@ -5,8 +5,7 @@
 && make -C Python \
 && make -C Python check \
 && make -C Python install \
-&& for i in Python/examples/*.py ; do echo $i && /usr/bin/python2 $i || break -1 ; done \
-&& grep -v QuantLib requirements.txt > Python/notebooks/requirements.txt \
-&& pip install -r Python/notebooks/requirements.txt \
-&& for i in Python/notebooks/*.py ; do echo "$i" && /usr/bin/python2 "$i" || break -1 ; done
+&& grep -v QuantLib requirements.txt > Python/examples/requirements.txt \
+&& pip install -r Python/examples/requirements.txt \
+&& for i in Python/examples/*.py ; do echo "$i" && /usr/bin/python2 "$i" || break -1 ; done
