+++ Python/examples/README.md
@@ -1,6 +1,6 @@
 # QuantLib Python examples
 
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/lballabio/QuantLib-SWIG/binder?filepath=Python%2Fnotebooks)
+[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/lballabio/QuantLib-SWIG/binder?filepath=Python%2Fexamples)
 
 This directory contains a number of examples of using QuantLib from
 Python.  They can also run as Jupyter notebooks by means of
@@ -8,7 +8,7 @@ Python.  They can also run as Jupyter notebooks by means of
 
 You can try them without installing QuantLib on your computer; they
 are available on [Binder](https://mybinder.org/) at
-[https://mybinder.org/v2/gh/lballabio/QuantLib-SWIG/binder?filepath=Python%2Fnotebooks](https://mybinder.org/v2/gh/lballabio/QuantLib-SWIG/binder?filepath=Python%2Fnotebooks)
+[https://mybinder.org/v2/gh/lballabio/QuantLib-SWIG/binder?filepath=Python%2Fexamples](https://mybinder.org/v2/gh/lballabio/QuantLib-SWIG/binder?filepath=Python%2Fexamples)
 (or you can click the "Launch Binder" badge at the top of this file).
 
 If you want to run these examples locally, you'll need the modules listed in the
