+++ Python/examples/european-option.py
@@ -1,56 +1,72 @@
-# Copyright (C) 2004, 2005, 2006, 2007 StatPro Italia srl
+# ---
+# jupyter:
+#   jupytext:
+#     formats: md,py:percent
+#     text_representation:
+#       extension: .py
+#       format_name: percent
+#       format_version: '1.3'
+#       jupytext_version: 1.4.2
+#   kernelspec:
+#     display_name: Python 3
+#     language: python
+#     name: python3
+# ---
+
+# %% [markdown]
+# # European options
 #
-# This file is part of QuantLib, a free-software/open-source library
-# for financial quantitative analysts and developers - http://quantlib.org/
+# Copyright (&copy;) 2004, 2005, 2006, 2007, 2020 StatPro Italia srl
+#
+# This file is part of QuantLib, a free-software/open-source library for financial quantitative analysts and developers - https://www.quantlib.org/
 #
 # QuantLib is free software: you can redistribute it and/or modify it under the
 # terms of the QuantLib license.  You should have received a copy of the
 # license along with this program; if not, please email
 # <quantlib-dev@lists.sf.net>. The license is also available online at
-# <http://quantlib.org/license.shtml>.
+# <https://www.quantlib.org/license.shtml>.
 #
 # This program is distributed in the hope that it will be useful, but WITHOUT
 # ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 # FOR A PARTICULAR PURPOSE.  See the license for more details.
 
+# %%
 import QuantLib as ql
+import pandas as pd
 
-# global data
+# %% [markdown]
+# ### Global parameters
+
+# %%
 todaysDate = ql.Date(15, ql.May, 1998)
 ql.Settings.instance().evaluationDate = todaysDate
-settlementDate = ql.Date(17, ql.May, 1998)
-riskFreeRate = ql.FlatForward(settlementDate, 0.05, ql.Actual365Fixed())
 
-# option parameters
+# %%
+is_interactive = 'get_ipython' in globals()
+
+# %% [markdown]
+# ### Option construction
+
+# %%
 exercise = ql.EuropeanExercise(ql.Date(17, ql.May, 1999))
 payoff = ql.PlainVanillaPayoff(ql.Option.Call, 8.0)
 
-# market data
+# %%
+option = ql.VanillaOption(payoff, exercise)
+
+# %% [markdown]
+# ### Market data
+
+# %%
 underlying = ql.SimpleQuote(7.0)
-volatility = ql.BlackConstantVol(settlementDate, ql.TARGET(), 0.10, ql.Actual365Fixed())
-dividendYield = ql.FlatForward(settlementDate, 0.05, ql.Actual365Fixed())
+dividendYield = ql.FlatForward(todaysDate, 0.05, ql.Actual365Fixed())
+volatility = ql.BlackConstantVol(todaysDate, ql.TARGET(), 0.10, ql.Actual365Fixed())
+riskFreeRate = ql.FlatForward(todaysDate, 0.05, ql.Actual365Fixed())
 
-# report
-header = " |".join(["%17s" % tag for tag in ["method", "value", "estimated error", "actual error"]])
-print("")
-print(header)
-print("-" * len(header))
+# %% [markdown]
+# ### Processes and models
 
-refValue = None
-
-
-def report(method, x, dx=None):
-    e = "%.4f" % abs(x - refValue)
-    x = "%.5f" % x
-    if dx:
-        dx = "%.4f" % dx
-    else:
-        dx = "n/a"
-    print(" |".join(["%17s" % y for y in [method, x, dx, e]]))
-
-
-# good to go
-
+# %%
 process = ql.BlackScholesMertonProcess(
     ql.QuoteHandle(underlying),
     ql.YieldTermStructureHandle(dividendYield),
@@ -58,6 +74,7 @@ process = ql.BlackScholesMertonProcess(
     ql.BlackVolTermStructureHandle(volatility),
 )
 
+# %%
 hestonProcess = ql.HestonProcess(
     ql.YieldTermStructureHandle(riskFreeRate),
     ql.YieldTermStructureHandle(dividendYield),
@@ -70,67 +87,108 @@ hestonProcess = ql.HestonProcess(
 )
 hestonModel = ql.HestonModel(hestonProcess)
 
-option = ql.VanillaOption(payoff, exercise)
+# %% [markdown]
+# ### Pricing
 
-# method: analytic
+# %% [markdown]
+# We'll collect tuples of method name, option value, estimated error, and discrepancy from the analytic formula.
+
+# %%
+results = []
+
+# %% [markdown]
+# #### analytic formula
+
+# %%
 option.setPricingEngine(ql.AnalyticEuropeanEngine(process))
 value = option.NPV()
 refValue = value
-report("analytic", value)
 
-# method: Heston semi-analytic
+results.append(('Analytic', value, None, None))
+
+# %% [markdown]
+# #### Heston semi-analytic formula
+
+# %%
 option.setPricingEngine(ql.AnalyticHestonEngine(hestonModel))
-report("Heston analytic", option.NPV())
+value = option.NPV()
 
-# method: Heston COS method
+results.append(('Heston analytic', value, None, abs(value - refValue)))
+
+# %% [markdown]
+# #### Heston COS method
+
+# %%
 option.setPricingEngine(ql.COSHestonEngine(hestonModel))
-report("Heston COS Method", option.NPV())
+value = option.NPV()
 
-# method: Heston Exponentially-Fitted Gauss-Laguerre quadrature rule
-option.setPricingEngine(ql.ExponentialFittingHestonEngine(hestonModel))
-report("Heston ExpFitting", option.NPV())
+results.append(('Heston COS', value, None, abs(value - refValue)))
 
-# method: integral
+# %% [markdown]
+# #### Integral method
+
+# %%
 option.setPricingEngine(ql.IntegralEngine(process))
-report("integral", option.NPV())
+value = option.NPV()
 
-# method: finite differences
+results.append(('Integral', value, None, abs(value - refValue)))
+
+# %% [markdown]
+# #### Finite-difference method
+
+# %%
 timeSteps = 801
 gridPoints = 800
 
+# %%
 option.setPricingEngine(ql.FdBlackScholesVanillaEngine(process, timeSteps, gridPoints))
-report("finite diff.", option.NPV())
+value = option.NPV()
 
-# method: binomial
+results.append(('Finite diff.', value, None, abs(value - refValue)))
+
+# %% [markdown]
+# #### Binomial method
+
+# %%
 timeSteps = 801
 
-option.setPricingEngine(ql.BinomialVanillaEngine(process, "JR", timeSteps))
-report("binomial (JR)", option.NPV())
+# %%
+for tree in ["JR", "CRR", "EQP", "Trigeorgis", "Tian", "LR", "Joshi4"]:
+    option.setPricingEngine(ql.BinomialVanillaEngine(process, tree, timeSteps))
+    value = option.NPV()
 
-option.setPricingEngine(ql.BinomialVanillaEngine(process, "CRR", timeSteps))
-report("binomial (CRR)", option.NPV())
+    results.append(('Binomial (%s)' % tree, value, None, abs(value - refValue)))
 
-option.setPricingEngine(ql.BinomialVanillaEngine(process, "EQP", timeSteps))
-report("binomial (EQP)", option.NPV())
+# %% [markdown]
+# #### Monte Carlo method
 
-option.setPricingEngine(ql.BinomialVanillaEngine(process, "Trigeorgis", timeSteps))
-report("bin. (Trigeorgis)", option.NPV())
+# %%
+option.setPricingEngine(ql.MCEuropeanEngine(process, "pseudorandom", timeSteps=1,
+                                            requiredTolerance=0.02, seed=42))
+value = option.NPV()
 
-option.setPricingEngine(ql.BinomialVanillaEngine(process, "Tian", timeSteps))
-report("binomial (Tian)", option.NPV())
+results.append(("Monte Carlo (pseudo-random)", value, option.errorEstimate(), abs(value - refValue)))
 
-option.setPricingEngine(ql.BinomialVanillaEngine(process, "LR", timeSteps))
-report("binomial (LR)", option.NPV())
+# %%
+option.setPricingEngine(ql.MCEuropeanEngine(process, "lowdiscrepancy", timeSteps=1,
+                                            requiredSamples=32768))
+value = option.NPV()
 
-option.setPricingEngine(ql.BinomialVanillaEngine(process, "Joshi4", timeSteps))
-report("binomial (Joshi)", option.NPV())
+results.append(("Monte Carlo (low-discrepancy)", value, None, abs(value - refValue)))
 
-# method: finite differences
-# not yet implemented
+# %% [markdown]
+# #### Results
 
-# method: Monte Carlo
-option.setPricingEngine(ql.MCEuropeanEngine(process, "pseudorandom", timeSteps=1, requiredTolerance=0.02, seed=42))
-report("MC (crude)", option.NPV(), option.errorEstimate())
+# %%
+df = pd.DataFrame(results,
+                  columns=["Method", "Option value", "Error estimate", "Actual error"])
 
-option.setPricingEngine(ql.MCEuropeanEngine(process, "lowdiscrepancy", timeSteps=1, requiredSamples=32768))
-report("MC (Sobol)", option.NPV())
+# %%
+df.style.hide_index()
+
+# %% [markdown]
+# The following displays the results when this is run as a Python script (in which case the cell above is not displayed).
+
+# %%
+if not is_interactive:
+    print(df)
